+++ CHANGELOG.md
@@ -1,5 +1,16 @@
 # Changelog
 
+## [1.12.3-beta.0](https://github.com/Seungwoo321/genai-commit/compare/v1.12.2...v1.12.3-beta.0) (2026-04-26)
+
+### Refactoring
+
+* **ui:** change action prompt from expand to list type ([db31b18](https://github.com/Seungwoo321/genai-commit/commit/db31b18083919d46d3db6da6063038854c491733))
+## [1.12.2-beta.0](https://github.com/Seungwoo321/genai-commit/compare/v1.12.2...v1.12.3-beta.0) (2026-04-25)
+
+### Bug Fixes
+
+* **ui:** bind hotkeys to action prompt using inquirer expand ([46626fc](https://github.com/Seungwoo321/genai-commit/commit/46626fc6ab7538c19f722d17076cb2786e306827))
+
 ## [1.12.2](https://github.com/Seungwoo321/genai-commit/compare/v1.12.1...v1.12.2) (2026-04-25)
 
 ### Bug Fixes